fix: reference to editor only functions
